updated contracts to be upgradeable and use latest LZ packages, also added deploy scripts
