+++ plume-oft-adapter-migration/contracts/ethereum/OrbitERC20OFTAdapterUpgradeable.sol
@@ -1,13 +1,14 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.22;
 
+import { OFTAdapterUpgradeable } from "@layerzerolabs/oft-evm-upgradeable/contracts/oft/OFTAdapterUpgradeable.sol";
+
 import { IERC20 } from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
 import { SafeERC20 } from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import { OFTAdapter } from "@layerzerolabs/lz-evm-oapp-v2/contracts/oft/OFTAdapter.sol";
 
 import { IERC20Bridge } from "./bridge/IERC20Bridge.sol";
 
-contract OrbitERC20OFTAdapter is OFTAdapter {
+contract OrbitERC20OFTAdapterUpgradeable is OFTAdapterUpgradeable {
     using SafeERC20 for IERC20;
 
     IERC20Bridge private immutable bridge;
@@ -15,13 +16,19 @@ contract OrbitERC20OFTAdapter is OFTAdapter {
     constructor(
         address _token,
         address _layerZeroEndpoint,
-        address _owner,
         IERC20Bridge _bridge
-    ) OFTAdapter(_token, _layerZeroEndpoint, _owner) {
+    ) OFTAdapterUpgradeable(_token, _layerZeroEndpoint) {
         bridge = _bridge;
+        _disableInitializers();
+    }
+
+    function initialize(address _delegate) public initializer {
+        __OFTAdapter_init(_delegate);
+        __Ownable_init(_delegate);
     }
 
     function _debit(
+        address /*_from*/,
         uint256 _amountLD,
         uint256 _minAmountLD,
         uint32 _dstEid
