+++ plume-oft-adapter-migration/contracts/ethereum/libraries/Error.sol
