+++ plume-oft-adapter-migration/contracts/ethereum/mocks/BridgeStub.sol
