+++ plume-oft-adapter-migration/contracts/ethereum/mocks/InboxStub.sol
