--- plume-oft-adapter-migration/deploy/MyOFTAdapterUpgradeable.ts
@@ -1,39 +1,0 @@
-import { type DeployFunction } from 'hardhat-deploy/types'
-
-import { EndpointId, endpointIdToNetwork } from '@layerzerolabs/lz-definitions'
-import { getDeploymentAddressAndAbi } from '@layerzerolabs/lz-evm-sdk-v2'
-
-const contractName = 'MyOFTAdapterUpgradeable'
-
-const deploy: DeployFunction = async (hre) => {
-    const { deploy } = hre.deployments
-    const signer = (await hre.ethers.getSigners())[0]
-    console.log(`deploying ${contractName} on network: ${hre.network.name} with ${signer.address}`)
-
-    const eid = hre.network.config.eid as EndpointId
-    const lzNetworkName = endpointIdToNetwork(eid)
-
-    const { address } = getDeploymentAddressAndAbi(lzNetworkName, 'EndpointV2')
-
-    await deploy(contractName, {
-        from: signer.address,
-        args: ['0x', address], // replace '0x' with the address of the ERC-20 token
-        log: true,
-        waitConfirmations: 1,
-        skipIfAlreadyDeployed: false,
-        proxy: {
-            proxyContract: 'OpenZeppelinTransparentProxy',
-            owner: signer.address,
-            execute: {
-                init: {
-                    methodName: 'initialize',
-                    args: [signer.address],
-                },
-            },
-        },
-    })
-}
-
-deploy.tags = [contractName]
-
-export default deploy
